# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Klinik Gigi API",
        "description": "Dokumentasi API untuk sistem Klinik Gigi",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Default Server"
        }
    ],
    "paths": {
        "/api/jadwal-dokter": {
            "get": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "List jadwal dokter",
                "operationId": "7470f02a32814270bb2d266dfc5e12e3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "dokter_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "Buat jadwal dokter",
                "operationId": "8b476020845183c15d79cd209a183eea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jadwal-dokter/{jadwal_dokter}": {
            "get": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "Detail jadwal dokter",
                "operationId": "49bc9870ade45a93e412e44ce08b222b",
                "parameters": [
                    {
                        "name": "jadwal_dokter",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "Update jadwal dokter",
                "operationId": "2243acbe8cb728d789a896ba560b194e",
                "parameters": [
                    {
                        "name": "jadwal_dokter",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "Hapus jadwal dokter",
                "operationId": "b927b5ed56c85e2d9d2557fdb7f4cd73",
                "parameters": [
                    {
                        "name": "jadwal_dokter",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jadwal-dokter/dokter/{dokter_id}": {
            "get": {
                "tags": [
                    "JadwalDokter"
                ],
                "summary": "Jadwal per dokter (jika ada kolom dokter_id)",
                "operationId": "5934284fd5a0f11507274e725453a348",
                "parameters": [
                    {
                        "name": "dokter_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifikasis": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "List notifikasi untuk user login",
                "operationId": "c4e0c3fc9410a20a9a00618850942d48",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifikasis/{notifikasi}": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Detail notifikasi",
                "operationId": "92f443f8d2a75dea4229f11f5a4d926d",
                "parameters": [
                    {
                        "name": "notifikasi",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Hapus notifikasi",
                "operationId": "47e1d559f623676062d9f54fd7903141",
                "parameters": [
                    {
                        "name": "notifikasi",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifikasis/{notifikasi}/mark-read": {
            "post": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Tandai satu notifikasi sebagai sudah dibaca (jika ada kolom read_at / is_read)",
                "operationId": "56831895d00302457cdb41be93c06ad1",
                "parameters": [
                    {
                        "name": "notifikasi",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifikasis/mark-all-read": {
            "post": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Tandai semua notifikasi user sebagai sudah dibaca",
                "operationId": "890cbe8e346451c8ba82f43cc8a74583",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifikasis/unread-count": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Jumlah notifikasi belum dibaca",
                "operationId": "7d217e4145814e454be0230283134a14",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}